Lazy load user dropdown content
